+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## ****Unreleased****
 
+#### Fixed
+- Single comma operand should be treated as no operands during validation
+
 ## [1.15.0](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.14.0...1.15.0) (2024-09-20)
 
 #### Added