+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## ****Unreleased****
+## [1.15.1](https://github.com/eclipse-che4z/che-che4z-lsp-for-hlasm/compare/1.15.0...1.15.1) (2024-11-22)
 
 #### Fixed
 - Single comma operand should be treated as no operands during validation